QML DOM: Allow compatibility with MSVC2019 builds on MSVC2022
When the DOM library is compiled with MSVC2019, we can't use our
function_ref implementation, as it runs into a compiler bug. That does
however mean that MSVC 2019 and 2022 builds are not ABI compatible,
which is an issue for QDS, as they want to use a Qt build against MSVC
2019 with a QDS build on MSVC 2022.

Solve this by allowing users to set a QML_DOM_MSVC2019_COMPAT macro,
which forces the API to use std::function, and thus to stay compatible
with the ABI resulting from a MSVC2019 build.

Task-number: QTBUG-127761
Change-Id: Icc10ad324657fcf66260298e2f643205522b03ed
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>