QQuickNativeMenuItem: check window count before accessing the first window
The sync() function was written under the assumption that
the application had at least one QWindow which could be used
for reading the current DPR. But during tear-down, this might
not be the case, and can result in a crash (found from running
the SpreadSheet example).

This patch will check that we have a non-empty window list before
accessing the first element.

Pick-to: 6.8
Change-Id: I75b00de1c259a2ee30d88ff17e880da80e40947e
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>