qmlls: ignore duplicates and missing files when watching files
Sort the results from QQmlCodeModel::fileNamesToWatch and remove
duplicates to avoid searching for the same files multiple times.

Track files that do not exist in the source paths (for example the
headers from Qt Quick) and ignore them for the next run of
findFilePathsFromFileNames().

This allows qmlls to be much faster, for example when using completions
on the todo qtdoc demo where qmlls tried to search 600 non-existing
files, including files with empty file names and duplicates) in the
source folder.

Pick-to: 6.8
Fixes: QTBUG-126576
Change-Id: I67a0aa84fb4e9286760a68ae9766823ec165936a
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>