quicklint: disable controls-attached-property-reuse category by default
It seems that this category is just not ready to be enabled by default,
so disable it.

Tracking the work to re-enable it at QTBUG-127812.

Fix a test to enable the category before testing for the warning.

Pick-to: 6.7
Fixes: QTBUG-127650
Change-Id: I6144e4b55e3bea803b14d82f6f0d291676d4394e
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit b0fd8ef05e4107c9fe21ba14f7f1de7a9a60e779)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>