FluentWinUI3 Style: Fix HighlightedText palette role color
Highlighted text represents the color of selected text and should be
white on both light and dark mode. We were incorrectly using this role
for the button text color for checked and highlighted buttons, which
does not correspond to always light text. Fix this for Button and
ToolButton as well as set this palette role to white color.

Pick-to: 6.8
Change-Id: Ia79d3e86cc038525e4c566d368f418ee5c56f89f
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>