PopupWindow: Don't consume keyboard events when focus: false
If the QGuiApplication forwards a key event to an active popup, and the
popup consumes the event, it will not be sent to the active focus window.

The ComboBox relies on the popup not getting focus, when dealing with
key events, and won't work as long as the PopupWindow gets and consumes
the keyboard events.

It makes sense to only have popups consume events when it has explicitly
been given keyboard focus.

In addition, unblacklist popupWindowFocus after some tweaks. It's an
important test which verifies various things related to focus.
Unfortunately, due to unknown reasons, it still doesn't pass on QNX.

Task-number: QTBUG-121363
Pick-to: 6.8
Change-Id: I459c91c5db5c392e997aa996f03ecdda093a544a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>