FluentWinUI3 Style: Override platform theme's palette with style's one
On Windows platforms we were using the platform theme's palette to
construct the FluentWinUI3 style's palette with the motivation that
the colors as gotten from the system's APIs in qwindowstheme would be
the correct ones for the style. This is not the case as these APIs
represent the colors for old Windows10 platforms and do not correspond
with the modern WinUI3 colors.
That is why we need to override many of these palette roles in the
FluentWinUI3's style palette. These colors need to be hardcoded even
for Windows platforms as there are no APIs that we can use to retrieve
them. The color entries are taken from the official Figma files for the
style. The accent color is an exception and that we can safely retrieve
from the system APIs.
As a drive-by, resolve the style's palette based on the platform theme
for all platforms, not only on Windows, as we want to use the system's
accent color no matter the platform.
Partially reverts 8516ddf690440cd33e3332e8cb2d049ed0032954.

Change-Id: I360fc30bd04fac9b04a8ffe0e99791349ad365a6
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 15e63383bcc0bba21f8c71c720fec4cf4ea5ab9f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>