qmlls: find and use resource files from .qt/rcc
Search for resource files in .qt/rcc instead of only looking in
the .rcc folder, and update the resource file mapper in
setLoadPaths to actually use the resource files.

Amends d0fcb75aab1b5a91260fb378d761257e1f2e4787 to:
 a) work on more sophisticated project structures with
    "prefer"-directives
 b) use resource files for all qmlls modules (previously only the
    linting module was using resource files)

Also move the methods to find .qrc files into qqmljsutils_p.h and add a
new test file.

This patch fixes the spurious warning about SettingsData being not found
and the missing completion suggestion for the SettingsData type in
qml/popups/AdminDialog.qml on Ekke's project attached at QTBUG-126504.
The build folder structure was mimicked in sophisticatedBuildFolder for
the test.

Also change DomEnvironment::semanticAnalysis() to return a value instead
of a reference, and rename SemanticAnalysis::setLoadPaths to
SemanticAnalysis::updateLoadPaths because it will set the loadpaths on
all copies of semanticAnalysis instead of just setting it on the current
one.

Pick-to: 6.8 6.7
Task-number: QTBUG-127661
Task-number: QTBUG-126680
Change-Id: I47299711ced6cb952c4d3c1893495cb17a8eec7f
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>