Android: Support multiple display densities in qtabstractlistmodel_kotlin
Added dimens.xml files for different display densities from where
the button margins, textsizes and paddings are fetched according to
the current display density.

Added corner radiuses for the buttons, to make the corners look the
same on different densities.

Added top and bottom margins and padding for the buttons, to make them
look the same on different densities.

Task-number: QTBUG-127848
Pick-to: 6.8
Change-Id: I9e7332af7abe71a4df3d2555803839017aec0896
Reviewed-by: Tinja Paavoseppä <tinja.paavoseppa@qt.io>