Compiler: Don't stop after first diagnostic
The type propagator currently skips all subsequent instructions as soon
as an error occurs. We want to be able to provide more than one warning
at the time within the same binding/function.

Therefore, stop skipping all instructions on the first error. The output
accumulator is set to a var after an error to avoid leaving it in an
invalid state and to be able to reuse later code. Some extra checks were
also added to avoid crashes.

Fixes: QTBUG-127624
Change-Id: Ifc0389767a067181c6e3ba4d5c3fc1a0597c29ef
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>