Reset handler state when dragged item dropped outside the drop area
The handler state had been reset only after the release state. This
causes the handler to be inconsistent when the dragged item is dropped
outside the corresponding drop area. It can even lead to a crash (as
mentioned in the bug report) when the next drop happens over the same
location.

This patch resets the state of the handler whenever a drop happens (as
this causes the drag event loop to exit).

Fixes: QTBUG-127865
Pick-to: 6.8
Change-Id: I14cac162df8af8b7b6b9eed99631e29bada410d1
Reviewed-by: MohammadHossein Qanbari <mohammad.qanbari@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>