Dialogs: Fall back to non-native dialog if native dialog can't be shown
A native dialog helper might not be able to show a dialog due to the
dialog options not being possible to represent with the native dialog.

If so the helper returns false from show(). We were not handling this
case, so failing to show a native dialog would result in no dialog
shown.

We now fall back to the non-native dialog by destroying the native
helper and using the non-native dialog helper explicitly.

Fixes: QTBUG-124124
Pick-to: 6.8 6.5
Change-Id: If50c63e419ae7b13e9317cd2cbbcaa02fe13bbd5
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>