tests: Fix target name overlap
In the build of the `tests/auto/qml/qqmlmoduleplugin/pluginWrongCase`
library, the target name overlapped with that of the library in
`tests/auto/qml/qqmlmoduleplugin/plugin`.

At least in the case of the (officially unsupported) MSVS generator this
resulted in an invalid build configuration (generated a solution file
containing two projects with the same name).

Change-Id: I050f6f854b734632003fda92b34666faf2137f50
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>