tools: Update target name of app in `tools/qml`
Target was previously named `qml`, which overlaps with the target name
in `src/qml`. At least in the case of the (officially unsupported) MSVS
generator, this resulted in an invalid build configuration (generated a
solution file containing two projects with the same name).

The target name was used to determine the name of the app's executable
file (e.g. `qml.exe` in Windows). This is now achieved through the
`OUTPUT_NAME` target property.

Change-Id: Ifbccceab11c8ad23e7f1b7a8cc7570d66cb6941f
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>