Update text components with color fonts when dpr changes
There was already a hook in text components to update
when the renderType was set to Text.NativeRendering. This
is necessary because the text is rendered into a glyph
cache at a given size and scaling it further will give
pixelation artifacts.

However, there are cases where the NativeRendering
backend is used implicitly. In particular this happens
for color fonts, so when changing the dpr of a screen
(or when moving the window to a screen with a different
dpr) we would get pixelated emojis in some cases.

The fix is to detect in the text node whether any
glyphs use the NativeRendering backend and if they
do, we update whenever the dpr changes.

Fixes: QTBUG-121449
Pick-to: 6.7 6.5
Change-Id: I9ad62d792e495d4439715fe87acfd87dd1833c67
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit 23ae81c84de8bb30ba2d779c603a5b3aded2a417)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>