Menu: Activate Menu Item When Mnemonic Key is Pressed, Same as Widgets
When a menu is opened, its menu items aren't triggered by pressing their
mnemonic key using the Alt key. Also, aren't triggered even when their
mnemonic key is pressed without the Alt key. This behavior is different
from the widgets menu behavior. When the widgets menu is opened its
actions can be activated by pressing their mnemonic key without holding
the Alt key.

When a key is pressed on QQuickMenu and no modifier is held, it loops
over its items to find the item whose mnemonic key matches the key. If
the item is found, the click() method will be called.

The test case opens the menu and sends the press and release events with
the mnemonic key of each menu item, and checks that they are triggered.
When the item is checkable, the test verifies that the checkedChanged()
signal is also emitted.

Task-number: QTBUG-96630
Pick-to: 6.8 6.7 6.5
Change-Id: I99f27c6306e14375b45d0630e4f9047e33fa64fb
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>