Fix invalid memory access during highlight creation in the item view
The creation of highlight item in the item views can go recursive
leading to invalid reference in the item thats been used to keep
track of the highlighted item.

The highlighted item has been reset with the newly created
highlighted item due to the geometry change in the item views,
which happened while creating the highlighted item itself. This is
like creation in a recursive loop. Since the highlighted item
referred by the tracked item, has been reset (in the second call to
the createHighlight()), the existing reference in the tracked item
automatically becomes invalid.                      

This patch adds validation to avoid recreating the highlighted item
in the item views if the creation request is already in progress.

Fixes: QTBUG-127455
Pick-to: 6.7 6.5
Change-Id: Ib4908d84e24c773e2b43a12c00ba23259896e171
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 81c95fb650a15261b9ba08e1dfe99f5f03fc3ad2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>