CMake: Exclude QML *_tooling targets from i18n
For QML modules, we create *_tooling targets with source files assigned,
which are located in the build directory. These sources are copies of
files from the source directory and already belong to the actual QML
module's target.

When using qt_add_translations' automatic target collection, both
targets contribute to the inputs of lupdate, resulting in duplicate
source locations in .ts files.

We now mark *_tooling targets as excluded from translation to avoid
these duplicates.

Implementation note: we don't need to set QT_EXCLUDE_FROM_TRANSLATION on
the utility target for CMake < 3.19, because utility targets do not
contribute to the input of lupdate.

Pick-to: 6.7 6.8
Fixes: QTBUG-124764
Change-Id: If5a8a19ff5503a6ea7d62a1f05957c7f1e8f315e
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>