DragHandler: Don't change QEventPoint accepted state in touch events
QQuickDragHandler::handlePointerEventImpl() accepts the event to begin
with. Since ca7cdd71ee33f0d77eb6bf1367d2532e26155cb2 we meant to stop
event propagation for mouse events, at least to prevent Flickable from
receiving the event via direct delivery after it has already filtered.
But back then, QEvent::setAccepted() did not automatically accept the
touchpoints; now it does (a questionable choice in Qt 6 that accept()
and setAccepted(true) do not do the same thing).

Now we leave QEventPoint:isAccepted state alone if it's a touch event,
so the workaround to stop propagation of mouse events is the only case
where it needs to be changed. And update the comment: since
3073a81e90c8d835dfccccf8b3a080a93ed0d2af Flickable handles touch events
directly, it doesn't depend on touch->mouse synthesis.

Fixes: QTBUG-124731
Pick-to: 6.8 6.7
Change-Id: I833a4d28f708e27cf01b108da1b886807725dc06
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>