Add a Gauge how-to
Gauge was part of Qt Quick Extras, but its implementation used Loader
and was very heavy. This patch implements it using C++ to avoid that.

As Gauge is quite complex and has a lot of possiblities for how users
could style it, this patch offers an example that can be used as a
starting point, and customized using a familiar delegate-based API.

The goal is to provide a set of C++ and QML files that users can simply
copy into their projects.

Fixes: QTBUG-111770
Change-Id: I389058296d65063e9ab6f6e484c59eb526b058b0
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>