Fix center in overlay behavior for QQuickPopupWindow
QQuickPopupPositioner does not correctly handle the case when there is
an anchor to center in the overlay, and the overlay is not the parent.
Also, remove some long ternary operations to make the code clearer and
fix some rounding issues when calculating the centers.

Task-number: QTBUG-126553
Change-Id: Ic186fef536f0a32b5e3c658ad7e64a81c1919cad
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit 1d59ba0a607c56819127484e5b27270d29e86af3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>