2D Renderer: Make sure cachedMirroredPixmap is never dirty when painting
When using the software context for Qt Quick scene graph rendering, it
was possible that Items using "layer.enabled: = true" would not be up to
date if these sub-passes were updated while the tree using them was not
visible (Due to update()/updatePaintNode() not being called). If the
cachedMirrorPixmap doesn't get updated, then the previous image will be
used when the item is painted, which is incorrect, and the update()
method will not be called again until the next time the item is dirty.

This change makes it where if the cache is still dirty when we start
painting, we update it then, because spending a bit more time renderer
is still better than rendering the wrong thing.

Fixes: QTBUG-114984
Pick-to: 6.7 6.5 5.15
Change-Id: I70f9f6d1dfd46d6870a2bee2ae72294e8982b776
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit 29a52df82397fd8a00308f7e34b08059d6affc91)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>