doc, QQuickHeaderView: ensure example snippet works
As it stood, the qdoc marker in the example code left the
final curly brace out of the snippet. A copy paste of this
example would therefore not compile.

This patch will make sure the brace is a part of the snippet
shown in the docs.

Task-number: QTBUG-123595
Pick-to: 6.7 6.5 6.2
Change-Id: I68bfac0cfc3dcecd487bfb69de07c317f787552e
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
(cherry picked from commit 3968ef8650da91c8e66d1651f7891793ac3b5d74)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>