Explicitly use popupType: Popup.Item in tst_qquickapplicationwindow
The test focusAfterPopupClosed(), checks to see that the delivery agent
causes correct updates to which item has active focus in a given scene.
However, when using popup windows, a popup's popupItem is moved into a
different scene, and the delivery agent won't respond the same.

We already have tests in tst_qquickpopup, to verify similar behavior
with popup windows.

Set the value to be explicitly Popup.Item, in case we change the default
value in the future.

Task-number: QTBUG-126553
Pick-to: 6.8
Change-Id: Ie1b2319adafc2b1a616d6b153fee49ac472a972c
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>