Make tst_QQuickMenu::contextMenuKeyboard work with popup windows
The test would fail if popup windows were enabled.
By making the test data driven, and fixing it so that it works with and
without popup windows enabled, we'll ensure that regressions are
detected, regardless of popupType.

It's also needed if we decide to change the default behavior in the
future.

Task-number: QTBUG-126553
Change-Id: Ia3da9c810a313c2b5e847f70d70ac36ec74d672d
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 3601648ba911330d41b4504d80827b40b3650834)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>