Consider Popup Windows when matching QQuickShortcuts
When the same key sequence is shared between multiple QQuickShortcuts,
it must be decided which QQuickShortcut takes precedence, i.e. is
'matched'.  If a QQuickItem owning a QQuickShortcut is blocked by a
modal popup, it should no longer be considered to be matched. Modify the
current implementation that checks for modal popups to also consider
popup windows.

Task-number: QTBUG-126553
Pick-to: 6.8
Change-Id: Ic6997756e4b9c924c6c3108407b6e5cdf5ec0f8a
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>