Use QWindow::mapToGlobal() directly
Fix issue where popup menus were incorrectly positioned
in cases where a QWindow spans multiple screens, and
the screens have different scale factors.

QWindow::mapToGlobal() must be called with the actual local
point to be mapped in order to produce correct results
for the corner cases where the point is on a different
screen than window (0,0).

Remove the indirection via QTransform and move the implementation
to QQuickItem::mapToGlobal(), and similar for mapFromGlobal().

Fixes: QTBUG-120011
Pick-to: 6.8
Change-Id: Id9182a0d55337b2771996e145905c961121ba8f8
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>