TextInput/TextField: Deselect text on tap or click
We now use the same approach as in QQuickTextControlPrivate:
on press, set hadSelectionOnMousePress according to whether text
is already selected or not; on release, if hadSelectionOnMousePress
is true, we should move the cursor (and thus deselect text).

Some history:

Commit 650342de792e0ab37ce8bac8ccde21ab9b96b2c9 disabled direct text
selection on touchscreens, because mobile platforms provide text
selection handles, and otherwise the user's finger is usually too
large relative to the font size to select text precisely, as one does
with a mouse. The test added there is now amended: if text was already
selected on press, we do expect tapping to move the cursor and deselect
the text.

Commit ac3f1eace1102543008a0ceded541e40ff495368 allowed importing an
old Qt Quick version to allow tapping to deselect.

Commit d6ddfa3d4ca3fbbaddc2acd7ebe3f8296fc68b68 fixed another bug:
text would lose selection when releasing afer a long press. But text
could not be deselected by touch. One way was to tap on something else
to make the text field lose focus.

Fixes: QTBUG-127440
Task-number: QTBUG-10684
Task-number: QTBUG-38934
Task-number: QTBUG-101205
Task-number: QTBUG-111504
Task-number: QTBUG-116606
Pick-to: 6.8 6.7 6.5
Change-Id: I5001216fc30266fcce3af413ecfa091bfd99beed
Reviewed-by: Ville Voutilainen <ville.voutilainen@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>