Android: Change member names and related docs to .loadContent rename
As the QtQuickView.loadComponent() became QtQuickView.loadContent()
and the QtQmlComponent renamed to QtQuickViewContent, the member names
in examples and related documentation content and documentation links
need to reflect the change, where required.

This change includes renamings, where they are applicable. There are
some cases that the word "component" is a better match, eg. when the
sentence is referring to signals or properties which are directly
talking about the QML component but not the QtQuickViewContent.

Task-number: QTBUG-127084
Pick-to: 6.8
Change-Id: I33ff259d3dfc97fb6b776dd676c19c3eea0842e4
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>