Force Popup::test_resetSize to use `popupType: Popup.Item`
It is not possible to reset the size of a QWindow, but it is possible to
reset the size of an QQuickItem. This test tests if resetting the size
of a QQuickPopup is successful, therefore, it does not need to be run
for popup windows.

Task-number: QTBUG-126553
Change-Id: I6a7aab8ae0e665d9d2074fa472a8e4ec7e114b41
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit 7d244701d0e0eae103195fc051098b24dd26f739)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>