Update Popup::test_xyBindingLoop to consider popupType: Popup.Window
The test Popup::test_xyBindingLoop sets up a binding loop between a
popup and its parent's x and y properties. It then checks that the
expected x and y values are set. The resulting x and y values are
non-integer, something which windows do not support. Therefore, it is
needed to first take the floor of the result before checking that the
expected x and y values have been set.

Task-number: QTBUG-126553
Pick-to: 6.8
Change-Id: I4d8a62b8b9bee0ab53c1675b16902b20edd8b349
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>