add util to map .qml from source to build folder
Add a helper method qmlSourcePathFromBuildPath in the QQmlJSUtils static
class that maps QML files from the build folder to their source folder,
and vice-versa.

Uses a heuristic for QML modules that do not use "prefer" qmldirs for
each folder that contains QML files.

It will be used in a later commit to translate build folder paths,
encountered during "go to definiton" in qmlls, to source folder paths.

Pick-to: 6.8 6.7
Task-number: QTBUG-127661
Change-Id: I81f4b75a008d773835affb622af7f964b753e5ac
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>