FluentWinUI3: Support button and accent palette roles for Button/ToolButton
We want to be able to support custom palette.accent and palette.button
colors for the Button/ToolButton. To do that, use QML instead of images
for the button background and factor it out in reusable component
ButtonBackground.
Remove previously generated images for the button/toolbutton backgorunds
but keep using geometry information generated from the stylegenerator.

Pick-to: 6.8
Change-Id: I6f195b32142d06c96833ac3d9184da9f8bbd22ba
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>