Make tst_QQuickMenu::subMenuPosition consider popupType
Since we're debating whether or not to change the default popupType.
We'll need to fix some tests that break when the popupType is Window.
subMenuPosition is one of those tests that break.

Fix the test so that it passes, regardless of the default popupType.
Since it's data driven, we can run it both with and without popup
windows.

Since flipping is implemented in a different test, it doesn't quite make
sense to test it in this function.

Mirroring doesn't currently work when using popup windows.

Task-number: QTBUG-126553
Pick-to: 6.8
Change-Id: Ie07d0ab867a90724f10ef7e5acbee60e314c7f9f
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>