Make tst_qquickmenubar::menuPosition data driven based on popupType
Many tests were written with the assumption that every item would be in
the same scene. But since we introduced popup windows in 6.8, we should
run relevant tests using `popupType: Popup.Item` and
`popupType: Popup.Window`.

For now, we're limiting testing of `popupType: Popup.Window` to Windows
and macOS, with the goal to increase the number of platforms over time.

Task-number: QTBUG-126553
Pick-to: 6.8
Change-Id: I05ffcd5c9a3f83856402745317462777daa4d63d
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>