VectorImage: Allow root items to be path containers
When the generators were refactored to add special-handling of the
root node, the path container logic from the structure node handling
was not included there, causing some SVGs to get a lot of Shape items
instead of combining them. This re-introduces the logic by adding
a Shape item inside the root item if we see that we can make a root
path container. (This is slightly different from structure nodes,
where the structure node itself can be the Shape item. Since the
root node has special requirements, we keep it and just add a
new Shape level inside.)

Pick-to: 6.8
Fixes: QTBUG-126716
Change-Id: If05fd38bad08749cf5c4b338ead104aa01672e49
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>