QQuickHeaderView: for explicitly assigned models, document that data() will be used
The detailed documentation for HeaderView says that
if you assign a QAIM to the model property, the models
headerData() will be used. But in the documentation for the
model property itself, it says that data() will be used
(except when using a QAbstractTableModel).
Both cannot be true.

The actual behavior is that we use data() when the model is
assigned explicit to the model property, and headerData()
when the syncView's model is used. It's debatable whether
or not we should have always used headerData(), but changing
that now could cause regressions. Such a change would
also stop e.g QML ListModel from being used as a model,
which is also an argument for keeping it as-is.

This patch will therefore fix up the documention to be
consistent, and to inform that data() will be used.

Fixes: QTBUG-123595
Pick-to: 6.2
Change-Id: Id64977a3db5f00b4269f0e5ffb6a31e5ad607725
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit f4bc018fa9e57bdd72556473040622acac508ea7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 4e4e8328c6ab7f80d868d57f98fbd08f54cbd97d)
(cherry picked from commit 1b42a3405944dc5890fa70eac7b72e6430f3f8c7)