QQuickHeaderView: Remove the use of a QTransposeProxyModel
When assigning a list model to a HorizontalHeaderView, the layout
needs to be transposed because TableView (which HeaderView inherits)
will otherwise layout in a column-major order. But there are two
ways this can be done: in the model or in the view. And as it
stood, HeaderView would actually do both.

It's preferred to transpose the view instead of the model. This
is faster and doesn't require the use of a proxy model. And it
also supports transposing models other than QAIM (e.g a JavaScript
arrays). The reason QTransposeProxyModel was used in HeaderView in
the first place was (probably) because the support for transposing
TableView was implemented after HeaderView was implemented.

This patch will therefore remove the code in QQuickHeaderView
that tranposes the model. This will remove some complexity
from the code, and reduce memory overhead.

Note that transposing the view will now always be done
when needed from QQuickHeaderViewBasePrivate::syncModel().

Change-Id: Ic53eeaffcc947bcff76667a23dcfd390a917ba63
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>